Deleting Account functionality + Web version

+++ lib/frontend/pages/home/account/delete_account_page.dart
@@ -13,6 +13,7 @@ import 'package:secret_hitler/frontend/widgets/components/buttons/primary_elevat
 import 'package:secret_hitler/frontend/widgets/components/text/explaining_text.dart';
 import 'package:secret_hitler/frontend/widgets/components/useful_widgets/text_form_field.dart';
 import 'package:secret_hitler/frontend/widgets/header/header.dart';
+import 'package:secret_hitler/frontend/widgets/loading_spin.dart';
 
 class DeleteAccount extends ConsumerStatefulWidget {
   const DeleteAccount({super.key});
@@ -32,6 +33,8 @@ class _DeleteAccountState extends ConsumerState<DeleteAccount> {
   // Text field keys
   final GlobalKey<CustomTextFormFieldState> deleteAccountTextFieldKey = GlobalKey<CustomTextFormFieldState>();
 
+  bool _accountDeleted = false;
+
   void _goBack(BuildContext context) {
     Navigator.pop(context);
   }
@@ -39,7 +42,6 @@ class _DeleteAccountState extends ConsumerState<DeleteAccount> {
   @override
   Widget build(BuildContext context) {
     final userState = ref.read(userStateProvider);
-    final authApi = ref.read(authApiProvider);
     return PopScope(
       canPop: false,
       onPopInvokedWithResult: (didpop, _) async {
@@ -77,26 +79,39 @@ class _DeleteAccountState extends ConsumerState<DeleteAccount> {
               PrimaryElevatedButton(
                 text: AppLanguage.getLanguageData()['Delete'],
                 onPressed: () async {
+                  if (_accountDeleted) {return;}
+                  _accountDeleted = true;
                   String userInput = deleteAccountTextController.text.trim();
                   if (userInput == '') {
+                    LoadingSpin.openLoadingSpin(context);
                     deleteAccountTextFieldKey.currentState?.resetsErrors();
                     // LoadingSpin.openLoadingSpin(context);
                     bool response = await deleteUser(
                       userState.user!.$id,
-                      authApi.getClient(),
+                      ref,
                     );
+                    LoadingSpin.closeLoadingSpin(context);
                     if (!response) {
                       CustomSnackbar.showSnackbar(
                         AppLanguage.getLanguageData()['There was an error. Please try again!'],
                         Colors.red,
                         const Duration(seconds: 3),
                       );
+                      // Account deletion was successful
+                    } else {
+                      await CustomSnackbar.showSnackbar(
+                        AppLanguage.getLanguageData()['Your account has been successfully deleted!'],
+                        Colors.green,
+                        const Duration(seconds: 3),
+                      );
+                      _goBack(context);
                     }
                   } else {
                     deleteAccountTextFieldKey.currentState?.showError(
                         AppLanguage.getLanguageData()['Enter "DELETE"'],
                     );
                   }
+                  _accountDeleted = false;
                 },
               ),
             ],
